+++ routes/creneau.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 router.get("/", async (req,res) => {
     try {
@@ -35,46 +36,61 @@ router.get("/zone/:id", async (req,res) => {
     }
 })
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
-        const {debut,fin} = req.body
-        const creneau = await pool.query("insert into creneau (creneau_debut,creneau_fin) values ($1,$2) returning *",[debut,fin])
-        if (creneau.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {debut,fin} = req.body
+            const creneau = await pool.query("insert into creneau (creneau_debut,creneau_fin) values ($1,$2) returning *",[debut,fin])
+            if (creneau.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(creneau.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(creneau.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.put("/:id", async (req,res) => {
+router.put("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const {debut,fin} = req.body
-        const creneau = await pool.query("update creneau set creneau_debut = $2, creneau_fin = $3 where creneau_id = $1 returning *",[id,debut,fin])
-        if (creneau.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const {debut,fin} = req.body
+            const creneau = await pool.query("update creneau set creneau_debut = $2, creneau_fin = $3 where creneau_id = $1 returning *",[id,debut,fin])
+            if (creneau.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(creneau.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(creneau.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.delete("/:id", async (req,res) => {
+router.delete("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const creneau = await pool.query("delete from creneau where creneau_id = $1 returning *",[id])
-        if (creneau.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const creneau = await pool.query("delete from creneau where creneau_id = $1 returning *",[id])
+            if (creneau.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.status(200)
+            }
         }
         else {
-            return res.status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
