+++ routes/jeu.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 router.get("/", async (req,res) => {
     try {
@@ -25,46 +26,61 @@ router.get("/:id", async (req,res) => {
     }
 })
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
-        const {name,type} = req.body
-        const jeu = await pool.query("insert into jeu (jeu_name, jeu_type) values ($1, $2) returning *",[name,type])
-        if (jeu.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {name,type} = req.body
+            const jeu = await pool.query("insert into jeu (jeu_name, jeu_type) values ($1, $2) returning *",[name,type])
+            if (jeu.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(jeu.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(jeu.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.put("/:id", async (req,res) => {
+router.put("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const {name,type} = req.body
-        const jeu = await pool.query("update jeu set jeu_name = $2, jeu_type = $3 where jeu_id = $1 returning *",[id,name,type])
-        if (jeu.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const {name,type} = req.body
+            const jeu = await pool.query("update jeu set jeu_name = $2, jeu_type = $3 where jeu_id = $1 returning *",[id,name,type])
+            if (jeu.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(jeu.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(jeu.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.delete("/:id", async (req,res) => {
+router.delete("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const jeu = await pool.query("delete from jeu where jeu_id = $1 returning *",[id])
-        if (jeu.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const jeu = await pool.query("delete from jeu where jeu_id = $1 returning *",[id])
+            if (jeu.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.status(200)
+            }
         }
         else {
-            return res.status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
