+++ routes/travail.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 router.get("/", async (req,res) => {
     try {
@@ -25,46 +26,61 @@ router.get("/:id", async (req,res) => {
     }
 })
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
-        const {benevole,zone,creneau} = req.body
-        const travail = await pool.query("insert into travail (travail_benevole, travail_zone, travail_creneau) values ($1, $2, $3) returning *",[benevole,zone,creneau])
-        if (travail.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {benevole,zone,creneau} = req.body
+            const travail = await pool.query("insert into travail (travail_benevole, travail_zone, travail_creneau) values ($1, $2, $3) returning *",[benevole,zone,creneau])
+            if (travail.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(travail.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(travail.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.put("/:id", async (req,res) => {
+router.put("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const {benevole,zone,creneau} = req.body
-        const travail = await pool.query("update travail set travail_benevole = $2, travail_zone = $3, travail_creneau = $4 where travail_id = $1 returning *",[id,benevole,zone,creneau])
-        if (travail.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const {benevole,zone,creneau} = req.body
+            const travail = await pool.query("update travail set travail_benevole = $2, travail_zone = $3, travail_creneau = $4 where travail_id = $1 returning *",[id,benevole,zone,creneau])
+            if (travail.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(travail.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(travail.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.delete("/:id", async (req,res) => {
+router.delete("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const travail = await pool.query("delete from travail where travail_id = $1 returning *",[id])
-        if (travail.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const travail = await pool.query("delete from travail where travail_id = $1 returning *",[id])
+            if (travail.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.status(200)
+            }
         }
         else {
-            return res.status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
