+++ routes/zone.js
@@ -1,5 +1,6 @@
 const router = require("express").Router()
 const pool = require("../db")
+const auth = require("../utils/auth")
 
 router.get("/", async (req,res) => {
     try {
@@ -35,46 +36,61 @@ router.get("/creneau/:id", async (req,res) => {
     }
 })
 
-router.post("/", async (req,res) => {
+router.post("/", auth, async (req,res) => {
     try {
-        const {name} = req.body
-        const zone = await pool.query("insert into zone (zone_name) values ($1) returning *",[name])
-        if (zone.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {name} = req.body
+            const zone = await pool.query("insert into zone (zone_name) values ($1) returning *",[name])
+            if (zone.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(zone.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(zone.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.put("/:id", async (req,res) => {
+router.put("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const {name} = req.body
-        const zone = await pool.query("update zone set zone_name = $2 where zone_id = $1 returning *",[id,name])
-        if (zone.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const {name} = req.body
+            const zone = await pool.query("update zone set zone_name = $2 where zone_id = $1 returning *",[id,name])
+            if (zone.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.json(zone.rows[0]).status(200)
+            }
         }
         else {
-            return res.json(zone.rows[0]).status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
     }
 })
 
-router.delete("/:id", async (req,res) => {
+router.delete("/:id", auth, async (req,res) => {
     try {
-        const {id} = req.params
-        const zone = await pool.query("delete from zone where zone_id = $1 returning *",[id])
-        if (zone.rows.length === 0) {
-            return res.status(500)
+        if (req.role === "admin") {
+            const {id} = req.params
+            const zone = await pool.query("delete from zone where zone_id = $1 returning *",[id])
+            if (zone.rows.length === 0) {
+                return res.status(500)
+            }
+            else {
+                return res.status(200)
+            }
         }
         else {
-            return res.status(200)
+            return res.status(403).send("Not Authorized")
         }
     } catch (err) {
         console.error(err.message)
